Inline bytez.TerminateNewline into diffdoc; remove bytez package
The bytez package held a single exported function, TerminateNewline,
used only by diffdoc. A dedicated top-level package for one helper with
one consumer was unwarranted, so inline it as an unexported
terminateNewline in diffdoc and delete the package.

Behavior is unchanged. The test moves to an internal diffdoc test and
gains a no-alias case asserting that terminating an unterminated input
neither mutates it nor shares a backing array with the result.
